finished refactoring timed wwyds to use new command encapsulation structure as well as made improvements to normal wwyd and random wwyd commands
